+++ CHANGELOG.md
@@ -7,6 +7,14 @@ The format is based on Keep a Changelog and this project adheres to Semantic Ver
 - Keep a Changelog: https://keepachangelog.com/en/1.1.0/
 - Semantic Versioning: https://semver.org/spec/v2.0.0.html
 
+
+## [0.5.0] - 2025-11-20
+### Added
+- Add filters at the top of the left panel:
+  - Filter by discovered/undiscovered spells (radio buttons).
+  - Filter by mod (checkboxes for each mod present in the archive).
+
+
 ## [0.4.6] - 2025-11-19
 ### Added
 - Add "Easy layout" toggle in config, that will try to automatically adjust the right panel width.
@@ -16,6 +24,7 @@ The format is based on Keep a Changelog and this project adheres to Semantic Ver
 ### Fixed
 - Fix issue of books list not being shifted far enough on non-zero pages, causing a noticeable overlap of books between pages.
 
+
 ## [0.4.5] - 2025-11-17
 ### Fixed
 - Fix Spell books from other mods using ebwizardry Spell books, resulting in broken books and not being able to extract them correctly.
@@ -24,6 +33,7 @@ The format is based on Keep a Changelog and this project adheres to Semantic Ver
 ## [0.4.4] - 2025-11-17
 ### Fixed
 - Fix crash on dedicated servers.
+
 
 ## [0.4.2] - 2025-11-16
 ### Added